+++ src/bans/mod.rs
@@ -8,9 +8,10 @@ use semver::{Version, VersionReq};
 use std::cmp::Ordering;
 
 #[derive(Eq)]
+#[cfg_attr(test, derive(Debug))]
 pub struct KrateId {
-    name: String,
-    version: VersionReq,
+    pub(crate) name: String,
+    pub(crate) version: VersionReq,
 }
 
 impl Ord for KrateId {
@@ -40,26 +41,26 @@ fn binary_search<'a>(
 ) -> Result<(usize, &'a cfg::Skrate), usize> {
     let lowest = VersionReq::exact(&Version::new(0, 0, 0));
 
-    match arr.binary_search_by(|i| match i.item.name.cmp(&details.name) {
-        Ordering::Equal => i.item.version.cmp(&lowest),
+    match arr.binary_search_by(|i| match i.value.name.cmp(&details.name) {
+        Ordering::Equal => i.value.version.cmp(&lowest),
         o => o,
     }) {
         Ok(i) => Ok((i, &arr[i])),
         Err(i) => {
             // Backtrack 1 if the crate name matches, as, for instance, wildcards will be sorted
             // before the 0.0.0 version
-            let begin = if i > 0 && arr[i - 1].item.name == details.name {
+            let begin = if i > 0 && arr[i - 1].value.name == details.name {
                 i - 1
             } else {
                 i
             };
 
             for (j, krate) in arr[begin..].iter().enumerate() {
-                if krate.item.name != details.name {
+                if krate.value.name != details.name {
                     break;
                 }
 
-                if krate.item.version.matches(&details.version) {
+                if krate.value.version.matches(&details.version) {
                     return Ok((begin + j, krate));
                 }
             }
@@ -86,7 +87,7 @@ struct TreeSkipper {
 
 impl TreeSkipper {
     fn build(
-        skip_roots: Vec<toml::Spanned<TreeSkip>>,
+        skip_roots: Vec<crate::Spanned<TreeSkip>>,
         krates: &Krates,
         file_id: codespan::FileId,
         sender: crossbeam::channel::Sender<Pack>,
@@ -96,7 +97,7 @@ impl TreeSkipper {
         for ts in skip_roots {
             let num_roots = roots.len();
 
-            for krate in krates.search_matches(&ts.get_ref().id.name, &ts.get_ref().id.version) {
+            for krate in krates.search_matches(&ts.value.id.name, &ts.value.id.version) {
                 roots.push(Self::build_skip_root(ts.clone(), krate.0, krates));
             }
 
@@ -106,11 +107,7 @@ impl TreeSkipper {
                         Diagnostic::new(
                             Severity::Warning,
                             "skip tree root was not found in the dependency graph",
-                            Label::new(
-                                file_id,
-                                ts.start() as u32..ts.end() as u32,
-                                "no crate matched these criteria",
-                            ),
+                            Label::new(file_id, ts.span, "no crate matched these criteria"),
                         )
                         .into(),
                     )
@@ -122,12 +119,12 @@ impl TreeSkipper {
     }
 
     fn build_skip_root(
-        ts: toml::Spanned<TreeSkip>,
+        ts: crate::Spanned<TreeSkip>,
         krate_id: krates::NodeId,
         krates: &Krates,
     ) -> SkipRoot {
-        let span = ts.start() as u32..ts.end() as u32;
-        let ts = ts.into_inner();
+        let span = ts.span;
+        let ts = ts.value;
 
         let max_depth = ts.depth.unwrap_or(std::usize::MAX);
         let mut skip_crates = Vec::with_capacity(10);
@@ -492,7 +489,7 @@ mod test {
                     ..Default::default()
                 }
             )
-            .map(|(_, s)| &s.item.version)
+            .map(|(_, s)| &s.value.version)
             .unwrap(),
             &(VersionReq::parse("=0.3.1").unwrap())
         );
@@ -506,7 +503,7 @@ mod test {
                     ..Default::default()
                 }
             )
-            .map(|(_, s)| &s.item.version)
+            .map(|(_, s)| &s.value.version)
             .unwrap(),
             &(VersionReq::any())
         );
@@ -530,7 +527,7 @@ mod test {
                     ..Default::default()
                 }
             )
-            .map(|(_, s)| &s.item.version)
+            .map(|(_, s)| &s.value.version)
             .unwrap(),
             &(VersionReq::parse("=0.1.43").unwrap())
         );
@@ -544,7 +541,7 @@ mod test {
                     ..Default::default()
                 }
             )
-            .map(|(_, s)| &s.item.version)
+            .map(|(_, s)| &s.value.version)
             .unwrap(),
             &(VersionReq::parse("<0.1.42").unwrap())
         );
@@ -558,7 +555,7 @@ mod test {
                     ..Default::default()
                 }
             )
-            .map(|(_, s)| &s.item.version)
+            .map(|(_, s)| &s.value.version)
             .unwrap(),
             &(VersionReq::parse(">0.1.43").unwrap())
         );
@@ -572,7 +569,7 @@ mod test {
                     ..Default::default()
                 }
             )
-            .map(|(_, s)| &s.item.version)
+            .map(|(_, s)| &s.value.version)
             .unwrap(),
             &(VersionReq::parse("<0.1").unwrap())
         );
@@ -586,7 +583,7 @@ mod test {
                     ..Default::default()
                 }
             )
-            .map(|(_, s)| &s.item.version)
+            .map(|(_, s)| &s.value.version)
             .unwrap(),
             &(VersionReq::parse("<0.3").unwrap())
         );